Move some more stuff to common
