+++ src/reformatters/noaa/gfs/forecast/reformat_internals.py
@@ -1,23 +1,22 @@
 import concurrent
 import os
-import warnings
 from collections.abc import Generator, Iterable, Sequence
 from concurrent.futures import Future, ProcessPoolExecutor, ThreadPoolExecutor
-from contextlib import closing, contextmanager
 from functools import partial
 from itertools import batched, groupby, product
-from multiprocessing.shared_memory import SharedMemory
 from pathlib import Path
 
-import numpy as np
 import pandas as pd
 import xarray as xr
 import zarr
 
-from reformatters.common.iterating import consume, shard_slice_indexers
+from reformatters.common.iterating import consume
 from reformatters.common.logging import get_logger
-from reformatters.common.types import ArrayFloat32
-from reformatters.common.zarr import create_data_array_and_template
+from reformatters.common.zarr import (
+    create_data_array_and_template,
+    create_shared_buffer,
+    write_shards,
+)
 from reformatters.noaa.gfs.forecast import template
 from reformatters.noaa.gfs.read_data import SourceFileCoords, download_file, read_into
 from reformatters.noaa.noaa_config_models import NOAADataVar
@@ -136,7 +135,6 @@ def get_max_lead_times(
     return max_lead_times
 
 
-# TODO: Basically common except for the data_var type
 def filter_coords_and_paths(
     data_var: NOAADataVar, coords_and_paths: CoordsAndPaths
 ) -> CoordsAndPaths:
@@ -151,7 +149,6 @@ def filter_coords_and_paths(
         return coords_and_paths
 
 
-# TODO: Basically common except for the data_var type
 def read_into_data_array(
     out: xr.DataArray,
     data_var: NOAADataVar,
@@ -188,66 +185,6 @@ def read_into_data_array(
 #             data_array.values,
 #             keep_mantissa_bits=keep_mantissa_bits,
 #         )
-
-
-# TODO: Probably move to common zarr module
-def write_shards(
-    data_array_template: xr.DataArray,
-    store: zarr.storage.FsspecStore | Path,
-    shared_buffer: SharedMemory,
-    cpu_process_executor: ProcessPoolExecutor,
-) -> None:
-    shard_indexers = tuple(shard_slice_indexers(data_array_template))
-    chunk_init_times_str = ", ".join(
-        data_array_template.init_time.dt.strftime("%Y-%m-%dT%H:%M").values
-    )
-    logger.info(
-        f"Writing {data_array_template.name} {chunk_init_times_str} in {len(shard_indexers)} shards"
-    )
-
-    # Use ProcessPoolExecutor for parallel writing of shards.
-    # Pass only a lightweight template and the name of the shared buffer
-    # to avoid ProcessPool pickling very large arrays.
-    consume(
-        cpu_process_executor.map(
-            partial(
-                write_shard_to_zarr,
-                data_array_template,
-                shared_buffer.name,
-                store,
-            ),
-            shard_indexers,
-        )
-    )
-
-
-# TODO: Probably move to common zarr module
-def write_shard_to_zarr(
-    data_array_template: xr.DataArray,
-    shared_buffer_name: str,
-    store: zarr.storage.FsspecStore | Path,
-    shard_indexer: tuple[slice, ...],
-) -> None:
-    """Write a shard of data to zarr storage using shared memory."""
-    with (
-        warnings.catch_warnings(),
-        closing(SharedMemory(name=shared_buffer_name)) as shared_memory,
-    ):
-        shared_array: ArrayFloat32 = np.ndarray(
-            data_array_template.shape,
-            dtype=data_array_template.dtype,
-            buffer=shared_memory.buf,
-        )
-
-        data_array = data_array_template.copy()
-        data_array.data = shared_array
-
-        warnings.filterwarnings(
-            "ignore",
-            message="In a future version of xarray decode_timedelta will default to False rather than None.",
-            category=FutureWarning,
-        )
-        data_array[shard_indexer].to_zarr(store, region="auto")  # type: ignore[call-overload]
 
 
 type ChunkCoordinates = Sequence[SourceFileCoords]
@@ -323,14 +260,3 @@ def get_download_var_group_futures(
         ] = data_vars
 
     return download_var_group_futures
-
-
-# TODO: Probably move to common zarr module
-@contextmanager
-def create_shared_buffer(size: int) -> Generator[SharedMemory, None, None]:
-    try:
-        shared_memory = SharedMemory(create=True, size=size)
-        yield shared_memory
-    finally:
-        shared_memory.close()
-        shared_memory.unlink()
